+++ IBS.DataAccess/Repository/Filpride/PurchaseOrderRepository.cs
@@ -48,7 +48,7 @@ namespace IBS.DataAccess.Repository.Filpride
 
             var lastSeries = lastPo.PurchaseOrderNo!;
             var numericPart = lastSeries.Substring(2);
-            var incrementedNumber = int.Parse(numericPart) + 1;
+            var incrementedNumber = long.Parse(numericPart) + 1;
 
             return lastSeries.Substring(0, 2) + incrementedNumber.ToString("D10");
         }
@@ -72,7 +72,7 @@ namespace IBS.DataAccess.Repository.Filpride
 
             var lastSeries = lastPo.PurchaseOrderNo!;
             var numericPart = lastSeries.Substring(3);
-            var incrementedNumber = int.Parse(numericPart) + 1;
+            var incrementedNumber = long.Parse(numericPart) + 1;
 
             return lastSeries.Substring(0, 3) + incrementedNumber.ToString("D9");
         }
